Refactoring all 'Bactools' namespaces to 'abacat'
