+++ .gitignore
@@ -1,5 +1,5 @@
 __pycache__
-bactools.egg-info/
+abacat.egg-info/
 dist/
 build/
 accessions*
