+++ abacat/config.py
@@ -6,7 +6,7 @@ CONFIG = {
         "COG": "/Users/viniWS/Bio/db/COG/prot2003-2014.fa",
         "megares": "/Users/viniWS/Bio/db/megares_v1.01/megares_database_v1.01.fasta",
         "phenotyping": "/Users/viniWS/Bio/db/phenotyping/phenotyping.fasta",
-        "pathways": "/Users/viniWS/Bio/bactools/pipelines/pathways.json"
+        "pathways": "/Users/viniWS/Bio/abacat/pipelines/pathways.json"
     },
     "threads": int(os.cpu_count() / 2),
     "blast": {"evalue": 10 ** -20},
