+++ setup.py
@@ -1,10 +1,10 @@
 import setuptools
 
 setuptools.setup(
-    name="bactools",
+    name="abacat",
     version="0.0.1",
     author="Vini Salazar",
     author_email="viniws@gmail.com",
-    description="bactools - Bacterial Genome Curation Tools",
+    description="abacat - a bacterial annotation and curation toolkit",
     packages=setuptools.find_packages(),
 )
